Don't use len(errorsChan) as the number of errors to expect
The length of a buffered channel is the number of elements waiting in it.
We can have a race condition by which some of the goroutines have not yet sent
their results, and thus this len() is lower than 2*len(self.externalSinks),
and we just don't grab all the errors.

This situation also results in Store() returning before it's finished doing
its thing.